Moved fs2.Trace to fs2.internal.Trace
